Various edits & implemented rectangle overlap alg.
I worked it out in a physical notebook (thats new to me) and decided the
best solution is a corner detection case-by-case solver. There are many
convenient restrictions that let this be realistically solvable by mioa:
the rectangles are on a grid, both aligned on the same axes and discrete
in measure (single pixel precision).