Too much for one commit
I did a lot of testing with the slate design and functions (holy shit
css animations are based). The main feature of the slates is that they
are a preview of a webpage, which you drag to the top of the screen to
open. This is not necessarily intuitive when you consider the way links
usually function, so I need to include some kind of indication of this
functionality. I settled on a simple CSS animation which raises the
iframe by 5px while rounding the corners. More is needed, but this
animation does something for now.