A lot of editing and testing the slate css
I'm trying to come up with a system for navigating into the slates in an
interesting way, and my original idea of dragging them to the top center
of the screen seems unintuitive. I think there will be either a link in
the html of the iframe, or a link in the frame of the iframme somehow.