Modified newsticker code and added messages
I added some new static messages to the newsticker reel and I was
experimenting with dynamic ones (just a Math.random()) and I noticed
that after it went off the screen, it poked back out if the decimal part
had more digits. To fix this I modified the newsticker code to include
a "safety margin" (currently 10px) that it moves the items past the edge
of the screen. It works for now, but I might rewrite the newsticker code
to better acommodate more extremely dynamic messages.