Fixed the whole issue with the jsons not reading
I am bad at programming. I was checking my for loop counter against the
size of the array I JUST INITIALIZED meaning that it would ADD ONE (1)
ITEM AND CHECK THE SIZE AND BE LIKE "yup all done here" AND CLOSE THE
FOR LOOP. For reference, it should check the size of the array of slate
names, because that doesn't change.